--- Assets/Animations/test.txt.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0e92362a9672a7d4280834d27c0a917e
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
