--- Assets/Audio/test.txt.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: efb668d7c43032848be7685be8c7befe
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
