--- Assets/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d97bc1acf537dca48bbce9f1580cd2aa
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
