--- Assets/Editor/test.txt.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b3ad40011e72f4e41b36f8d231f23910
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
