--- Assets/NPC.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3311b32362efdec4db16f3b463d18262
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
