+++ Assets/Prefabs/Character.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 22ec3c52b34288949a89f79fb96cc1d8
-TextScriptImporter:
+guid: 05663ca9322dc554e9ab5c87cef45f07
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
