+++ Assets/Prefabs/Character/Player.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 1975863534}
   - component: {fileID: 6626503895012448124}
   - component: {fileID: 6626503895012448125}
+  - component: {fileID: 8270293506662018232}
   - component: {fileID: 341799687}
   m_Layer: 0
   m_Name: Player
@@ -103,6 +104,18 @@ CapsuleCollider2D:
   m_Offset: {x: 0, y: 0}
   m_Size: {x: 2.26, y: 4.89}
   m_Direction: 0
+--- !u!114 &8270293506662018232
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6626503895012448112}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: fe468ece471247c4eaa0435f694c65d7, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!50 &341799687
 Rigidbody2D:
   serializedVersion: 4
