+++ Assets/Prefabs/UI.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 96b97f9146a42fb418e49fbc173f1d1e
-TextScriptImporter:
+guid: 61f38a70dfe61634f9d33571266b3763
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
