--- Assets/Prefabs/test.txt.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 16c65737cd3af314d8259d74c47e9eb9
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
