--- Assets/Resources/test.txt.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3d286e3b8abafd44a9d7d56d15cdb239
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
