--- Assets/Scenes/LoadTest.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: cc03f4e254e09a24588f075b30deaf49
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
