+++ Assets/Scripts/Character.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: bfbf28613675de2489b0fcf717a1cf0e
-TextScriptImporter:
+guid: e9ccc8f6d9010dd4aa34ac9be58cea00
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
