--- Assets/Scripts/Conversati.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ecd758fa616f3db4f9b3959c1eab8f8b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
