--- Assets/Scripts/Conversation/ConversationData.asset
@@ -1,46 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5891681e936e3e943bf8fba4ac194268, type: 3}
-  m_Name: ConversationData
-  m_EditorClassIdentifier: 
-  items:
-  - type: 0
-    id: 0
-    targetID: 1
-    text: aaa
-    audio: {fileID: 0}
-    options:
-    - text: 
-      audio: {fileID: 0}
-      targetId: 0
-      enabled: 0
-    - text: 
-      audio: {fileID: 0}
-      targetId: 0
-      enabled: 0
-    - text: 
-      audio: {fileID: 0}
-      targetId: 0
-      enabled: 0
-    eventName: 
-  - type: 0
-    id: 1
-    targetID: 2
-    text: bbb
-    audio: {fileID: 0}
-    options:
-    - text: ccc
-      audio: {fileID: 0}
-      targetId: 0
-      enabled: 0
-    eventName: 
-  m_firstConversation: 0
