--- Assets/Scripts/Conversation/ConversationData.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a7b9c01bc5174284ab1517bff496a756
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
