--- Assets/Scripts/Conversation/NPCController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4ee225730a2b3074a8dc2c6326b49d84
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
