--- Assets/Scripts/MainMap/MoveEvent/Move_CircleWay_to_Room1.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 65a9fb4f973482c4f8b0b92fc1063399
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
