--- Assets/Scripts/MiniGameScripts/ItemManager.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-/// <summary>
-/// アイテムを管理する。
-/// </summary>
-public class ItemManager
-{
-    int[] Items;
-}
