--- Assets/Scripts/MiniGameScripts/ItemManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6c0d5532ed337434eae7486dfc0ef29b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
