--- Assets/Scripts/UniTaskTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 02c4099d5e51bce409db3a00f449a840
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
