+++ Assets/Scripts/Utility/SingletonMonoBehaviour.cs.meta
