--- Assets/Textures/test.txt.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 822ac321c7bcf734dbdf4b0842c1d30a
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
